fix(schema): ensure `SchemaGenerator` respects `skipTables` option when diffing a schema (#7040)
This updates the `prepareSchemaForComparison` method of
`SqlSchemaGenerator` to pass down the `skipTables` option from the
`schemaGenerator` config.

Prior to this change, this option was only being applied when generating
the `toSchema` but not `fromSchema`.

This posed an issue if you were not using snapshots as creating a fresh
schema as the target for `fromSchema` would not run through the same
"skip" logic that `getTargetSchema` would when generating the
`toSchema`.

This led to tables not being properly skipped and sql being generated
around these tables in the generated migrations.

Discord convo
[ref](https://discord.com/channels/1214904142443839538/1446299371816747018/1446539876802625628)

---------

Co-authored-by: Martin Adámek <banan23@gmail.com>